+++ .prettierrc.js
@@ -2,7 +2,7 @@ module.exports = {
 	singleQuote: true,
 	trailingComma: 'all',
 	bracketSpacing: true,
-	jsxBracketSameLine: false,
+	// jsxBracketSameLine: false,
 	useTabs: true,
 	semi: true,
 };